refactor: rename files and remove unused code
